requirements.txt
requests
pyyaml
pandas
matplotlib
pytest
python-dotenv
